# contracts/openapi-v1.yaml
openapi: 3.0.3
info:
  title: StayVista API v1
  version: 1.0.0
servers:
- url: http://localhost:8080
paths:
...
  /v1/admin/chat/curation/rules:
    get:
      summary: List chat curation rules
      parameters:
      - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeChatCurationRuleList'
    post:
      summary: Upsert chat curation rule
      parameters:
      - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCurationUpsertRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeChatCurationRule'
  /v1/admin/chat/curation/rules/{ruleId}:
    patch:
      summary: Update chat curation rule
      parameters:
      - $ref: '#/components/parameters/XRequestId'
      - name: ruleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCurationUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeChatCurationRule'
    delete:
      summary: Delete chat curation rule
      parameters:
      - $ref: '#/components/parameters/XRequestId'
      - name: ruleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
components:
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
...
  schemas:
    EnvelopeBase:
      type: object
      properties:
        request_id:
          type: string
      required:
      - request_id
...
    ChatCurationRule:
      type: object
      properties:
        rule_id:
          type: integer
          format: int64
        doc_id:
          type: string
        rule_type:
          type: string
          enum: [BLACKLIST, TOP_PICK]
        weight:
          type: integer
        enabled:
          type: boolean
        updated_at:
          type: string
      required:
      - rule_id
      - doc_id
      - rule_type
      - weight
      - enabled
      - updated_at
    ChatCurationUpsertRequest:
      type: object
      properties:
        doc_id:
          type: string
        rule_type:
          type: string
          enum: [BLACKLIST, TOP_PICK]
        weight:
          type: integer
          minimum: 1
          maximum: 500
        enabled:
          type: boolean
      required:
      - doc_id
      - rule_type
    ChatCurationUpdateRequest:
      type: object
      properties:
        weight:
          type: integer
          minimum: 1
          maximum: 500
        enabled:
          type: boolean
    ChatCurationRuleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChatCurationRule'
      required:
      - items
...
    EnvelopeChatCurationRule:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/ChatCurationRule'
        required:
        - data
    EnvelopeChatCurationRuleList:
      allOf:
      - $ref: '#/components/schemas/EnvelopeBase'
      - type: object
        properties:
          data:
            $ref: '#/components/schemas/ChatCurationRuleList'
        required:
        - data